backend/learning-analytics-module/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
lightgbm==4.1.0
joblib==1.3.2
Werkzeug==3.0.1
gunicorn==21.2.0